+++ dc-model/src/main/java/de/digitalcollections/model/identifiable/Identifiable.java
@@ -16,11 +16,7 @@ import de.digitalcollections.model.text.StructuredContent;
 import de.digitalcollections.model.text.contentblock.ContentBlock;
 import de.digitalcollections.model.text.contentblock.Paragraph;
 import de.digitalcollections.model.view.RenderingHintsPreviewImage;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Locale;
-import java.util.Objects;
-import java.util.Set;
+import java.util.*;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -418,9 +414,10 @@ public class Identifiable extends UniqueObject {
               .build());
     }
 
-    public B primaryLocalizedUrlAlias(String slug) {
+    public B primaryLocalizedUrlAlias(Locale targetLocale, String slug) {
       this.localizedUrlAliases =
-          new LocalizedUrlAliases(UrlAlias.builder().slug(slug).isPrimary().build());
+          new LocalizedUrlAliases(
+              UrlAlias.builder().targetLanguage(targetLocale).slug(slug).isPrimary().build());
       return self();
     }