+++ dc-model/src/test/java/de/digitalcollections/model/identifiable/entity/digitalobject/DigitalObjectTest.java
@@ -25,7 +25,7 @@ class DigitalObjectTest {
             .description(Locale.ENGLISH, "Example Image")
             .label(Locale.GERMAN, "Beispielbild")
             .label(Locale.ENGLISH, "Example Image")
-            .primaryLocalizedUrlAlias("blubb")
+            .primaryLocalizedUrlAlias(Locale.GERMAN, "blubb")
             .build();
     assertThat(digitalObject).isExactlyInstanceOf(DigitalObject.class);
     assertThat(digitalObject.getType()).isEqualTo(IdentifiableType.ENTITY);